added packaged structure, refactored docs and annotations a bit
