+++ eeg_models/types.py
@@ -1,0 +1,6 @@
+from typing import List, NewType, Optional
+
+import numpy as np
+
+
+npArray = NewType("nupmy array", np.ndarray)
